test: Replace fuzz_target_1 with fuzz_target_2
